emulate some classes from the jre
These can not be tested in development mode, but they are relatively simple, so it should be no problem.
Another caveat of this direct emulation: We cannot add public methods or fields that aren't preset in the Java Classpath.
But on the plus side, we don't need to fix the imports in JOSM code.
The emulation is minimalistic - more functionality should be added on demand.