add original JOSM classes (r4087)
For some classes, the package name is changed from
src.org.openstreetmap.josm.* to src.org.openstreetmap.josm_server.*.
These classes are executed on the server and not compiled to javascript.
GWT would have no problem putting these in a different source folder under the
original package name, but unfortunately the eclipse plugin gets confused, when
there is client and server code with the same package. It still seems to work,
but you'll get annoying warnings.